requirements.txt
Django==1.7.1
Pillow==2.6.1
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-ckeditor==5.0.3
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.4.8
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
